Task5

+++ src/main/java/service/StorageService.java
@@ -5,11 +5,12 @@ import api.service.IRequestService;
 import api.service.IStorageService;
 import dao.StorageDao;
 import exceptions.DaoException;
-import exceptions.ServiceException;
 import model.Book;
+import model.Order;
 import model.StatusBook;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import util.Config;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -45,7 +46,9 @@ public class StorageService implements IStorageService {
             book.setDateReceipt(localDate);
             storageDao.addBook(book);
             if (requestService.isRequest(book)) {
+                if (Boolean.parseBoolean(Config.configProperties("storageService.permissionChangeStatusRequest"))){
                 requestService.deleteRequest(requestService.getRequest(book));
+                }
             }
             return true;
         } catch (DaoException e) {
@@ -70,9 +73,10 @@ public class StorageService implements IStorageService {
     }
 
     @Override
-    public List<Book> printBookNotSellMoreSixMonth() {
+    public List<Book> BookNotSellMoreNmonth() {
+        int month = Integer.parseInt(Config.configProperties("storageService.month"));
         return storageDao.getBooks().stream()
-                .filter(book -> LocalDate.now().minusMonths(6).isAfter(book.getDateReceipt()))
+                .filter(book -> LocalDate.now().minusMonths(month).isAfter(book.getDateReceipt()))
                 .collect(Collectors.toList());
     }
 
@@ -81,4 +85,9 @@ public class StorageService implements IStorageService {
         return new ArrayList<>(storageDao.getBooks());
     }
 
+    @Override
+    public <T> void set(List<T> list) {
+        log.info("Десериализация Storage");
+        list.stream().map(e -> (Book) e).forEach(storageDao::addBook);
+    }
 }
